Add `demos/der_print_flexi.c`
