+++ demos/CMakeLists.txt
@@ -28,7 +28,7 @@ endif()
 #
 # Demos that are usable but only rarely make sense to be installed
 #
-# USEABLE_DEMOS  = aesgcm constants crypt openssh-privkey openssl-enc pem-info sizes timing
+# USEABLE_DEMOS  = aesgcm constants crypt der_print_flexi openssh-privkey openssl-enc pem-info sizes timing
 # -----------------------------------------------------------------------------
 
 if(BUILD_USABLE_DEMOS)
@@ -38,6 +38,7 @@ if(BUILD_USABLE_DEMOS)
         aesgcm
         constants
         crypt
+        der_print_flexi
         openssh-privkey
         openssl-enc
         pem-info
