+++ makefile.mingw
@@ -301,6 +301,8 @@ constants.exe: demos/constants.o $(LIBMAIN_S)
 	$(CC) demos/constants.o $(LIBMAIN_S) $(LTC_LDFLAGS) -o $@
 timing.exe: demos/timing.o $(LIBMAIN_S)
 	$(CC) demos/timing.o $(LIBMAIN_S) $(LTC_LDFLAGS) -o $@
+der_print_flexi.exe: demos/der_print_flexi.o $(LIBMAIN_S)
+	$(CC) demos/der_print_flexi.o $(LIBMAIN_S) $(LTC_LDFLAGS) -o $@
 
 #Tests
 test.exe: $(TOBJECTS) $(LIBMAIN_S)
