+++ makefile.unix
@@ -312,6 +312,8 @@ constants: demos/constants.o $(LIBMAIN_S)
 	$(CC) demos/constants.o $(LIBMAIN_S) $(LTC_LDFLAGS) -o $@
 timing: demos/timing.o $(LIBMAIN_S)
 	$(CC) demos/timing.o $(LIBMAIN_S) $(LTC_LDFLAGS) -o $@
+der_print_flexi: demos/der_print_flexi.o $(LIBMAIN_S)
+	$(CC) demos/der_print_flexi.o $(LIBMAIN_S) $(LTC_LDFLAGS) -o $@
 
 #Tests
 test: $(TOBJECTS) $(LIBMAIN_S)
